AI-Workout-Suggestion-main/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.23.2
pydantic==2.6.1
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai>=0.4.0
